add destory action reaction list and common action to mixins
